requirements.txt
pyftdi==0.52.9
proxmoxer==1.1.1
python-dotenv==0.17.1
requests==2.24.0
